+++ CHANGELOG.md
@@ -12,6 +12,8 @@
 
 * FIXED: whois.dk-hostmaster.dk parser should support status `Deactivated`.
 
+* FIXED: whois.domain-registry.nl parser should support status `Inactive`.
+
 * CHANGED: Changed .NAME to a formatted adapter to fetch additional properties.
 
 * CHANGED: whois.dns.pl now supports expires_on (GH-185). [Thanks @y3ti]